Improved selective enabling of undo buffers. We now can turn off undo buffering for in progress transactions. Added cheap hack to prune the Dtxn.Coordinator's queue when we're overloaded. Probably won't need this when we write the new CC-NUMA coordinator in august... There is a funky problem where this gets stuck when loading TPC-C in our experimental mode (but only TPC-C)... not sure what it could be
git-svn-id: https://database.cs.brown.edu/svn/hstore/branches/markov-branch@2188 7a8f3d5b-8c5f-44cb-9125-2414a256df87

+++ src/frontend/edu/brown/utils/ArgumentsParser.java
@@ -805,7 +805,7 @@ public class ArgumentsParser {
         }
         if (this.params.containsKey(PARAM_MARKOV_THRESHOLDS_VALUE)) {
             assert(this.catalog_db != null);
-            double defaultValue = this.getDoubleParam(PARAM_MARKOV_THRESHOLDS_VALUE);
+            float defaultValue = this.getDoubleParam(PARAM_MARKOV_THRESHOLDS_VALUE).floatValue();
             this.thresholds = new EstimationThresholds(defaultValue);
             LOG.debug("CREATED THRESHOLDS: " + this.thresholds);
